timeseries-statsmodels/requirements.txt
zenml[server]>=0.50.0
packaging
notebook
pandas_datareader
statsmodels
matplotlib